bootstrap: Install additional ACPI table with global NVS area address
The global nvs (GNVS) area address differs depending on device model,
BIOS settings and connected devices in the DSDT table. At the moment,
the DSDT override is necessary to fix various mistakes in the original
DSDT table. Overriding with a pre-compiled version is difficult because
the GNVS area address differs on each device.

Until now, this was handled using an out-of-tree patch in the kernel.
However, this does not work using mainline kernels. Additionally,
it requires building the ACPI DSDT table directly into the kernel.

With these changes, a new "OEM6" ACPI table is installed on each
boot - it contains only a single "NVSA" integer with the GNVS
address. This table is then loaded dynamically from the modified
DSDT table. This allows a static pre-compiled DSDT table without
the device specific offset.

Note: DSDT override could also be done in bootstrap but is
currently still handled by the kernel. That makes it possible
to enable/disable it using kernel parameters/additional initrds
for users that still want to run the stock Android system.